requirements.txt
requests
cloudshell-automation-api==2023.2.0.186356